Seats are now selectable and random occupied seats appear
Users can now select seats with their seat IDs saved in an array. Also, random occupied seats appear now as well. The only thing is that the selectedSeats array is not updated on the DOM so it does not accurately compare the seats and tickets on the seat selector popup