Seat Selector actually finalized and Checkout started
The seat selector pop-up functions correctly, with some testing still on the way. The Checkout page has been partially designed with an input form section and another section for the final ticket count and total price still being worked out. Also, the tab for some reason does not update to " - Checkout" even when directly swapping the words "Billing" and "Checkout".