Designed the Checkout page and Adding Correct Expiration years
The design of the checkout page is pretty much finished. The form is missing validation and checking if the text in fields is appropriate. And right now I am trying to populate the expiration year for the cards to have years from now to 25 years but it keeps repeating the useEffect hook that its in.